Update version number to match upcoming tag
The last time this gem was updated, the version number was not updated
in `lib/saml_idp_version.rb`. This PR updates the version to match the
upcoming `0.23.10-18f` tag that will be created after this PR is merged.

Also, I added a script to make it easier to create a new tag.